+++ src/lib/Scenes/MyProfile/__tests__/MyProfileEditForm.tests.tsx
@@ -1,46 +1,47 @@
-import { fireEvent, waitForElementToBeRemoved } from "@testing-library/react-native"
-import { MyProfileEditFormModalTestsQuery } from "__generated__/MyProfileEditFormModalTestsQuery.graphql"
+import { act, fireEvent, waitForElementToBeRemoved } from "@testing-library/react-native"
+import { MyProfileEditFormTestsQuery } from "__generated__/MyProfileEditFormTestsQuery.graphql"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { __globalStoreTestUtils__ } from "lib/store/GlobalStore"
+import { flushPromiseQueue } from "lib/tests/flushPromiseQueue"
 import { mockEnvironmentPayload } from "lib/tests/mockEnvironmentPayload"
 import { renderWithWrappersTL } from "lib/tests/renderWithWrappers"
 import React from "react"
 import { graphql, QueryRenderer } from "react-relay"
 import { createMockEnvironment } from "relay-test-utils"
-import { MyProfileEditFormModal } from "../MyProfileEditFormModal"
+import { MyProfileEditForm } from "../MyProfileEditForm"
 
 jest.unmock("react-relay")
+// jest.mock("lib/relay/createEnvironment", () => ({
+//   defaultEnvironment: require("relay-test-utils").createMockEnvironment(),
+// }))
 
-describe("MyProfileEditFormModal", () => {
-  let mockEnvironment: ReturnType<typeof createMockEnvironment>
+jest.mock("@react-navigation/native", () => {
+  const actualNav = jest.requireActual("@react-navigation/native")
+  return {
+    ...actualNav,
+    useNavigation: () => ({
+      navigate: jest.fn(),
+    }),
+  }
+})
 
-  beforeEach(() => {
-    mockEnvironment = createMockEnvironment()
-  })
+describe("MyProfileEditForm", () => {
+  const mockEnvironment = defaultEnvironment as ReturnType<typeof createMockEnvironment>
 
   const TestRenderer = () => (
-    <QueryRenderer<MyProfileEditFormModalTestsQuery>
+    <QueryRenderer<MyProfileEditFormTestsQuery>
       environment={mockEnvironment}
       query={graphql`
-        query MyProfileEditFormModalTestsQuery @relay_test_operation {
+        query MyProfileEditFormTestsQuery @relay_test_operation {
           me {
-            ...MyProfileEditFormModal_me @arguments(enableCollectorProfile: true)
+            ...MyProfileEditForm_me
           }
         }
       `}
       variables={{}}
       render={({ props }) => {
         if (props?.me) {
-          return (
-            <MyProfileEditFormModal
-              me={props.me}
-              visible
-              onDismiss={jest.fn}
-              setProfileIconLocally={jest.fn}
-              localImage={null}
-              refetchProfileIdentification={jest.fn}
-              relay={{ environment: mockEnvironment } as any}
-            />
-          )
+          return <MyProfileEditForm />
         }
         return null
       }}
@@ -60,7 +61,7 @@ describe("MyProfileEditFormModal", () => {
 
     describe("Email Verification", () => {
       describe("When the email is confirmed in Gravity", () => {
-        it("is shown as verified when it's verified in gravity", () => {
+        it("is shown as verified when it's verified in gravity", async () => {
           const { getByText } = renderWithWrappersTL(<TestRenderer />)
           mockEnvironmentPayload(mockEnvironment, {
             Me: () => ({
@@ -95,7 +96,9 @@ describe("MyProfileEditFormModal", () => {
           const VerifyYouEmailButton = getByTestId("verify-your-email")
           expect(VerifyYouEmailButton).toBeTruthy()
 
-          fireEvent(VerifyYouEmailButton, "onPress")
+          act(() => fireEvent(VerifyYouEmailButton, "onPress"))
+
+          await flushPromiseQueue()
 
           mockEnvironment.mock.resolveMostRecentOperation({
             data: {
